feat(memory): lower mem_move / mem_set on every backend (MEMORY-ACCESS M0)
`__builtin_mem_copy` reached the compiled lanes in Phase 3e; its two
siblings never did. `__builtin_mem_move` and `__builtin_mem_set` existed
only in the tree-walker, so any program using them died with "compiler
MVP cannot lower builtin yet: MemMove". Nothing in `core/std/*.t` called
either one -- which is why the gap survived, and part of why the stdlib
writes byte loops (213 `ptr_read` / `ptr_write` call sites) where a range
operation belongs. design-docs/MEMORY_ACCESS.md has the diagnosis.

Adds `InstKind::MemMove` / `InstKind::MemSet`, lowering for both, AOT
codegen against libc `memmove` / `memset`, and the matching IR VM host
methods.

`mem_set`'s fill value is now a `u8`, as `docs/language.md` and the AST
comment always said. The type checker's table said `u64`, the
tree-walker demanded `u64`, and nothing enforced either -- because
`visit_builtin_call` answers from the signature table *without visiting
the arguments at all*, so `arg_types` was decorative
(`__builtin_mem_set(p, undefined_name, 8u64)` type-checked and failed at
run time). The three `mem_*` builtins now check their arguments through
`check_memory_builtin_args`, which routes through `coerce_number_expr`
so a suffix-less literal and a character literal take the argument
position's type the way they do everywhere else.

Two consequences worth noting:

- `lower_builtin_call`'s catch-all arm is gone. These were the last two
  builtins without a lowering, so the match is exhaustive and a new
  `BuiltinFunction` now fails to compile there instead of failing at run
  time.
- `jit_heap.t` left `AOT_UNSUPPORTED` -- the two-directional skip list
  asked for it as soon as `mem_set` compiled.

+++ frontend/src/type_checker/visitor.rs
@@ -377,7 +377,11 @@ impl<'a> TypeCheckerVisitor<'a> {
             BuiltinFunctionSignature {
                 func: BuiltinFunction::MemSet,
                 arg_count: 3,
-                arg_types: vec![TypeDecl::Ptr, TypeDecl::UInt64, TypeDecl::UInt64],
+                // MEMORY-ACCESS M0: the fill value is one byte, as
+                // `docs/language.md` and the AST comment always said.
+                // It used to be `u64` here and each lane truncated it
+                // its own way.
+                arg_types: vec![TypeDecl::Ptr, TypeDecl::UInt8, TypeDecl::UInt64],
                 return_type: TypeDecl::Unit,
             },
             // Allocator handle builtins. The Allocator value itself is opaque at the
